+++ packages/typespec-rust/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## 0.28.1 (unreleased)
 
+### Features Added
+
+* Added support for custom date-time encoding `rfc3339-fixed-width`.
+
 ### Bugs Fixed
 
 * Fixed invalid function name for literal value `serde` helpers.